pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.7"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.17"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.0"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_path = "."
addopts = ['-p no:warnings']

[tool-ruff]
line-length = 79
extend-exclude= ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run     = 'fastapi dev fast_zero/app.py'

pre_test = 'task lint'
test     = 'pytest --cov=fast_zero -vv --ignore=fast_zero/settings.py'
post_test = 'coverage html'

lint    = 'ruff check . && ruff check . --diff'
format  = 'ruff check . --fix ruff format'


